Revert "++Refactor code structure and update dependencies, add Docker image"
