--- .dockerignore
@@ -1,5 +1,0 @@
-scripts/
-examples/
-data/
-.venv/
-render.yaml
