--- .env.example
@@ -1,2 +1,0 @@
-PDB_ZIP_PATH=data/3FTx/3FTx_pdb.zip
-DEFAULT_JSON_FILE_PATH=data/3FTx/3FTx_customized.json
