+++ script/download_foldcomp.sh
