+++ script/figures_script/merge_svg.py
