+++ script/prepare_json.py
